Extract libFunctional from libPlatform
