+++ src/Spectre.libFunctional.Tests/FilterTest.cpp
@@ -20,7 +20,7 @@ limitations under the License.
 #include <gtest/gtest.h>
 #include <gmock/gmock-matchers.h>
 #include "Spectre.libException/InconsistentArgumentSizesException.h"
-#include "Spectre.libPlatform/Filter.h"
+#include "Spectre.libFunctional/Filter.h"
 
 namespace
 {
