+++ src/Spectre.libFunctional.Tests/RangeTest.cpp
@@ -19,7 +19,7 @@ limitations under the License.
 
 #include <gtest/gtest.h>
 #include <gmock/gmock-matchers.h>
-#include "Spectre.libPlatform/Range.h"
+#include "Spectre.libFunctional/Range.h"
 
 namespace
 {
