+++ src/Spectre.libFunctional.Tests/packages.config
@@ -1,0 +1,5 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="gmock" version="1.7.0" targetFramework="native" />
+  <package id="Microsoft.Gsl" version="0.1.2.1" targetFramework="native" />
+</packages>
