+++ src/Spectre.libFunctional/Filter.h
