+++ src/Spectre.libFunctional/Find.cpp
