+++ src/Spectre.libFunctional/Range.h
