+++ src/Spectre.libFunctional/packages.config
@@ -1,0 +1,4 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="Microsoft.Gsl" version="0.1.2.1" targetFramework="native" />
+</packages>
