+++ src/Spectre.libPlatform/Spectre.libPlatform.vcxproj
@@ -73,20 +73,11 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
-    <ClInclude Include="Filter.h" />
-    <ClInclude Include="Find.h" />
     <ClInclude Include="Math.h" />
-    <ClInclude Include="Range.h" />
     <ClInclude Include="Statistics.h" />
-    <ClInclude Include="Transform.h" />
-    <ClInclude Include="ZeroStepException.h" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
-  </ItemGroup>
-  <ItemGroup>
-    <ClCompile Include="Find.cpp" />
-    <ClCompile Include="ZeroStepException.cpp" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
