+++ src/Spectre.libPlatform/Spectre.libPlatform.vcxproj.filters
@@ -13,12 +13,6 @@
       <UniqueIdentifier>{67DA6AB6-F800-4c08-8B7A-83BB121AAD01}</UniqueIdentifier>
       <Extensions>rc;ico;cur;bmp;dlg;rc2;rct;bin;rgs;gif;jpg;jpeg;jpe;resx;tiff;tif;png;wav;mfcribbon-ms</Extensions>
     </Filter>
-    <Filter Include="Header Files\Functional">
-      <UniqueIdentifier>{c08b655a-c602-433f-a1fd-9c23c8e34267}</UniqueIdentifier>
-    </Filter>
-    <Filter Include="Source Files\Functional">
-      <UniqueIdentifier>{42e3bea1-4acd-4800-9758-5ff237911f35}</UniqueIdentifier>
-    </Filter>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="Math.h">
@@ -27,31 +21,8 @@
     <ClInclude Include="Statistics.h">
       <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="Transform.h">
-      <Filter>Header Files\Functional</Filter>
-    </ClInclude>
-    <ClInclude Include="Range.h">
-      <Filter>Header Files\Functional</Filter>
-    </ClInclude>
-    <ClInclude Include="Filter.h">
-      <Filter>Header Files\Functional</Filter>
-    </ClInclude>
-    <ClInclude Include="Find.h">
-      <Filter>Header Files\Functional</Filter>
-    </ClInclude>
-    <ClInclude Include="ZeroStepException.h">
-      <Filter>Header Files\Functional</Filter>
-    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
-  </ItemGroup>
-  <ItemGroup>
-    <ClCompile Include="ZeroStepException.cpp">
-      <Filter>Source Files\Functional</Filter>
-    </ClCompile>
-    <ClCompile Include="Find.cpp">
-      <Filter>Source Files\Functional</Filter>
-    </ClCompile>
   </ItemGroup>
 </Project>
