+++ src/Spectre.sln
@@ -76,6 +76,7 @@ EndProject
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Spectre.libClassifier", "Spectre.libClassifier\Spectre.libClassifier.vcxproj", "{5419AEEB-98B4-4D79-9FCA-A794200073FC}"
 	ProjectSection(ProjectDependencies) = postProject
 		{7417BF00-028B-4797-B58A-6058CA338493} = {7417BF00-028B-4797-B58A-6058CA338493}
+		{7EB0161F-4E8A-4C72-BA18-31B554F411F2} = {7EB0161F-4E8A-4C72-BA18-31B554F411F2}
 		{B7DB309D-3ED6-4771-993D-A1CF3A6436D9} = {B7DB309D-3ED6-4771-993D-A1CF3A6436D9}
 	EndProjectSection
 EndProject
@@ -84,6 +85,7 @@ Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Spectre.libClassifier.Tests
 		{7417BF00-028B-4797-B58A-6058CA338493} = {7417BF00-028B-4797-B58A-6058CA338493}
 		{76542909-5FC8-405B-8E6D-D2E529162A7A} = {76542909-5FC8-405B-8E6D-D2E529162A7A}
 		{9786751A-6AE1-48D3-938D-4129D0399667} = {9786751A-6AE1-48D3-938D-4129D0399667}
+		{7EB0161F-4E8A-4C72-BA18-31B554F411F2} = {7EB0161F-4E8A-4C72-BA18-31B554F411F2}
 		{5419AEEB-98B4-4D79-9FCA-A794200073FC} = {5419AEEB-98B4-4D79-9FCA-A794200073FC}
 	EndProjectSection
 EndProject
@@ -111,6 +113,13 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Spectre.DataValidatorApp", 
 	EndProjectSection
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Spectre.DivikResultConverter", "Spectre.DivikResultConverter\Spectre.DivikResultConverter.csproj", "{74E04873-BB6F-4115-841D-50C162928B19}"
+EndProject
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Spectre.libFunctional", "Spectre.libFunctional\Spectre.libFunctional.vcxproj", "{7EB0161F-4E8A-4C72-BA18-31B554F411F2}"
+EndProject
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Spectre.libFunctional.Tests", "Spectre.libFunctional.Tests\Spectre.libFunctional.Tests.vcxproj", "{CF33269B-0876-4651-A8B9-2E32740C05C7}"
+	ProjectSection(ProjectDependencies) = postProject
+		{7EB0161F-4E8A-4C72-BA18-31B554F411F2} = {7EB0161F-4E8A-4C72-BA18-31B554F411F2}
+	EndProjectSection
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -250,6 +259,14 @@ Global
 		{74E04873-BB6F-4115-841D-50C162928B19}.Debug|x64.Build.0 = Debug|x64
 		{74E04873-BB6F-4115-841D-50C162928B19}.Release|x64.ActiveCfg = Release|x64
 		{74E04873-BB6F-4115-841D-50C162928B19}.Release|x64.Build.0 = Release|x64
+		{7EB0161F-4E8A-4C72-BA18-31B554F411F2}.Debug|x64.ActiveCfg = Debug|x64
+		{7EB0161F-4E8A-4C72-BA18-31B554F411F2}.Debug|x64.Build.0 = Debug|x64
+		{7EB0161F-4E8A-4C72-BA18-31B554F411F2}.Release|x64.ActiveCfg = Release|x64
+		{7EB0161F-4E8A-4C72-BA18-31B554F411F2}.Release|x64.Build.0 = Release|x64
+		{CF33269B-0876-4651-A8B9-2E32740C05C7}.Debug|x64.ActiveCfg = Debug|x64
+		{CF33269B-0876-4651-A8B9-2E32740C05C7}.Debug|x64.Build.0 = Debug|x64
+		{CF33269B-0876-4651-A8B9-2E32740C05C7}.Release|x64.ActiveCfg = Release|x64
+		{CF33269B-0876-4651-A8B9-2E32740C05C7}.Release|x64.Build.0 = Release|x64
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -274,6 +291,8 @@ Global
 		{81F10DE8-B9EF-400F-B3A7-57F2AFF35A14} = {7C77A75B-F387-4E4E-9C34-FA904927C7BF}
 		{3D932849-312C-473C-B74C-ABD1F11A0739} = {D3C36686-9667-48B2-A14B-FF46E6350A51}
 		{74E04873-BB6F-4115-841D-50C162928B19} = {D3C36686-9667-48B2-A14B-FF46E6350A51}
+		{7EB0161F-4E8A-4C72-BA18-31B554F411F2} = {7C77A75B-F387-4E4E-9C34-FA904927C7BF}
+		{CF33269B-0876-4651-A8B9-2E32740C05C7} = {7C77A75B-F387-4E4E-9C34-FA904927C7BF}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {ED57AF8B-7937-40C6-9764-D313AB08FEC9}
